requirements.txt
gunicorn
tensorflow-cpu==2.8.0
numpy
flask